Moving meme commands to new file meme.ts
