+++ commands/meme.ts
@@ -1,0 +1,17 @@
+const ngo = () => {
+	return `:middle_finger: 
+:no_entry_sign: :no_entry_sign: :no_entry_sign:
+NO COINERS GET OUT
+:poop: :poop: :poop:
+:thumbsdown:`
+}
+
+const spec = () => {
+	return `Warning!
+
+S P E C U L A T I O N
+
+Warning!`
+}
+
+export default { ngo, n: ngo, spec, s: spec}
